+++ client/src/main/java/com/influxdb/client/internal/AuthorizationsApiImpl.java
@@ -30,6 +30,7 @@ import javax.annotation.Nullable;
 import com.influxdb.Arguments;
 import com.influxdb.client.AuthorizationsApi;
 import com.influxdb.client.domain.Authorization;
+import com.influxdb.client.domain.AuthorizationPostRequest;
 import com.influxdb.client.domain.Authorizations;
 import com.influxdb.client.domain.Organization;
 import com.influxdb.client.domain.Permission;
@@ -88,7 +89,14 @@ final class AuthorizationsApiImpl extends AbstractRestClient implements Authoriz
 
         Arguments.checkNotNull(authorization, "Authorization is required");
 
-        Call<Authorization> call = service.postAuthorizations(authorization, null);
+        AuthorizationPostRequest request = new AuthorizationPostRequest();
+        request
+                .orgID(authorization.getOrgID())
+                .userID(authorization.getUserID())
+                .permissions(authorization.getPermissions())
+                .description(authorization.getDescription());
+
+        Call<Authorization> call = service.postAuthorizations(request, null);
 
         return execute(call);
     }